Contextualize database and separate in to modules (#486)
* database: Move verifier functions to their own file.

* database: Move some functions to operator module.

* database: Move header chain prover related things to their own module.

* database: Add watchtower module.

* database: Move rest of the functions to operator.

* database: Add file comments.

* Add execute_query_with_tx macro (#488)

* database: Add execute_query_with_tx macro.

* database: Use execute_query_with_tx.

* Update core/src/database/mod.rs


---------


* Update db function names, add tests and small improvements (#489)

* database: Add the new DatabaseTransaction type.

* database: Update comments and names of watchtower functions.

* database: Update verifier func comments/names.

* database: Update comments and add tests for operator.

* database: Remove Option from DatabaseTransaction type.

* database: Update header chain prover namings and docs.

---------
