+++ core/src/header_chain_prover/prover.rs
@@ -149,7 +149,7 @@ impl HeaderChainProver {
                     Ok(receipt) => {
                         prover
                             .db
-                            .save_block_proof(None, current_block_hash, receipt)
+                            .set_block_proof(None, current_block_hash, receipt)
                             .await
                             .unwrap();
                     }
@@ -270,10 +270,10 @@ mod tests {
         let hash =
             BlockHash::from_raw_hash(Hash::from_slice(&block_headers[1].prev_block_hash).unwrap());
         let header: Header = block_headers[0].clone().into();
-        let _ = prover.db.save_new_block(None, hash, header, 0).await; // TODO: Unwrapping this causes errors.
+        let _ = prover.db.set_new_block(None, hash, header, 0).await; // TODO: Unwrapping this causes errors.
         prover
             .db
-            .save_block_proof(None, hash, receipt.clone())
+            .set_block_proof(None, hash, receipt.clone())
             .await
             .unwrap();
         let database_receipt = prover.get_header_chain_proof(hash).await.unwrap();
@@ -300,13 +300,13 @@ mod tests {
         };
         prover
             .db
-            .save_new_block(None, hash, header, 0)
+            .set_new_block(None, hash, header, 0)
             .await
             .unwrap();
 
         prover
             .db
-            .save_block_proof(None, hash, receipt.clone())
+            .set_block_proof(None, hash, receipt.clone())
             .await
             .unwrap();
         let database_receipt2 = prover.get_header_chain_proof(hash).await.unwrap();
